+++ HISTORY.rst
@@ -18,6 +18,7 @@ History
 
 - Fixed issue #409
 - Sorted imports
+- Added downloading gridadmin file in the GeoPackage format #438
 
 
 3.4.2 (2023-02-08)